+++ packages/cervel/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @vercel/cervel
 
+## 0.1.25
+
+### Patch Changes
+
+- Updated dependencies [fdb6121]
+  - @vercel/backends@0.8.17
+
+## 0.1.24
+
+### Patch Changes
+
+- @vercel/backends@0.8.16
+
+## 0.1.23
+
+### Patch Changes
+
+- @vercel/backends@0.8.15
+
 ## 0.1.22
 
 ### Patch Changes